kgplus: extract ConstructExtendedSK to its own file
ConstructExtendedSK is dense secret-key R' math (interleaving
coefficient slots, NTT/IMForm conversions, basis extension). It has
nothing in common with TransmissionKeys other than both happening
during KG+ key generation, and its 70 lines were the bulk of keys.go.

Move it to a dedicated kgplus/extendedsk.go. keys.go now contains
exactly the package doc, the TransmissionKeys struct, and its
serialization methods — one cohesive concern per file.
